requirements.txt
pysimplegui
mysql-connector-python